Move episode_writer from the analysis package into evaluation
episode_writer is eval-time data production — it runs inside the eval loop, depends on
the metrics/evaluation machinery, and is called only by eval_runner. It has no coupling
to the analysis code (the analyzer consumes the JSONL purely as a format). Relocating it
beside its caller frees the sensitivity package of any pxr/sim dependency.
