+++ isaaclab_arena/evaluation/eval_runner.py
@@ -262,7 +262,7 @@ def main():
                     if episode_summary_enabled:
                         # Deferred import: episode_writer transitively touches pxr, so import
                         # after sim init — same pattern as policy_runner's compute_metrics.
-                        from isaaclab_arena.analysis.sensitivity.episode_writer import write_episode_summaries
+                        from isaaclab_arena.evaluation.episode_writer import write_episode_summaries
 
                         rows = write_episode_summaries(env, job, args_cli.episode_summary)
                         print(f"[INFO] Wrote {rows} episode summaries for job '{job.name}'")
